requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
sqlalchemy
pydantic
pydantic-settings
pytest
